+++ js/constants.js
@@ -1,0 +1,25 @@
+const descriptions = [
+  'Отличная получилась фотография',
+  'Чудесный день!',
+  'Небо зараза какое синее',
+  'Могу написать всё что угодно',
+  'Хороший кадр',
+  'Какой-то текст',
+  'Хорошего всем дня!',
+  'Вылазь из интернета и иди работай',
+  'Хехе',
+  'Мяу'
+];
+
+const messages = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const names = ['София', 'Виктория', 'Кирилл', 'Илья', 'Максим', 'Денис', 'Анастасия', 'Кира'];
+
+export {descriptions, messages, names};
